+++ compiler-plugin/src/test/kotlin/tech/mappie/testing/lists/ListPropertyPrimitiveToListPropertyPrimitiveTest.kt
@@ -7,27 +7,25 @@ import tech.mappie.testing.compilation.compile
 import tech.mappie.testing.loadObjectMappieClass
 import java.io.File
 
-class ObjectWithListObjectToObjectListObjectTest {
-    data class Input(val text: List<InnerInput>, val int: Int)
-    data class InnerInput(val value: String)
+class ListPropertyPrimitiveToListPropertyPrimitiveTest {
+    data class Input(val a: InnerInput)
+    data class InnerInput(val value: List<String>)
 
-    data class Output(val text: List<InnerOutput>, val int: Int)
-    data class InnerOutput(val value: String)
+    data class Output(val a: InnerOutput)
+    data class InnerOutput(val value: List<String>)
 
     @TempDir
     lateinit var directory: File
 
     @Test
-    fun `map nested list implicit should succeed`() {
+    fun `map object with nested list implicit should succeed`() {
         compile(directory) {
             file("Test.kt",
                 """
                 import tech.mappie.api.ObjectMappie
-                import tech.mappie.testing.lists.ObjectWithListObjectToObjectListObjectTest.*
+                import tech.mappie.testing.lists.ListPropertyPrimitiveToListPropertyPrimitiveTest.*
 
                 class Mapper : ObjectMappie<Input, Output>()
-
-                object InnerMapper : ObjectMappie<InnerInput, InnerOutput>()
                 """
             )
         } satisfies {
@@ -40,8 +38,8 @@ class ObjectWithListObjectToObjectListObjectTest {
                 .first()
                 .call()
 
-            assertThat(mapper.map(Input(listOf(InnerInput("first"), InnerInput("second")), 20)))
-                .isEqualTo(Output(listOf(InnerOutput("first"), InnerOutput("second")), 20))
+            assertThat(mapper.map(Input(InnerInput(listOf("first", "second")))))
+                .isEqualTo(Output(InnerOutput(listOf("first", "second"))))
         }
     }
 }